+++ Pandora/Paradigm/Primary/Transformer/Construction.hs
@@ -21,59 +21,59 @@ import Pandora.Pattern.Object.Ringoid ((*))
 import Pandora.Pattern.Object.Monoid (Monoid (zero))
 import Pandora.Paradigm.Controlflow.Joint.Schemes.TU (TU (TU))
 
-data Construction t a = Construction a (t :. Construction t := a)
+data Construction t a = Construct a (t :. Construction t := a)
 
 instance Covariant t => Covariant (Construction t) where
-	f <$> Construction x xs = Construction (f x) $ f <$$> xs
+	f <$> Construct x xs = Construct (f x) $ f <$$> xs
 
 instance Avoidable t => Pointable (Construction t) where
-	point x = Construction x empty
+	point x = Construct x empty
 
 instance Covariant t => Extractable (Construction t) where
-	extract (Construction x _) = x
+	extract (Construct x _) = x
 
 instance Applicative t => Applicative (Construction t) where
-	Construction f fs <*> Construction x xs = Construction (f x) $ fs <**> xs
+	Construct f fs <*> Construct x xs = Construct (f x) $ fs <**> xs
 
 instance Traversable t => Traversable (Construction t) where
-	Construction x xs ->> f = Construction <$> f x <*> xs ->>> f
+	Construct x xs ->> f = Construct <$> f x <*> xs ->>> f
 
 instance Alternative t => Bindable (Construction t) where
-	Construction x xs >>= f = case f x of Construction y ys -> Construction y $ ys <+> (>>= f) <$> xs
+	Construct x xs >>= f = case f x of Construct y ys -> Construct y $ ys <+> (>>= f) <$> xs
 
 instance Covariant t => Extendable (Construction t) where
-	x =>> f = Construction (f x) $ extend f <$> deconstruct x
+	x =>> f = Construct (f x) $ extend f <$> deconstruct x
 
 instance (Avoidable t, Alternative t) => Monad (Construction t) where
 
 instance Covariant t => Comonad (Construction t) where
 
 instance Lowerable Construction where
-	lower (Construction _ xs) = extract <$> xs
+	lower (Construct _ xs) = extract <$> xs
 
 instance (Setoid a, forall b . Setoid b => Setoid (t b)) => Setoid (Construction t a) where
-	Construction x xs == Construction y ys = (x == y) * (xs == ys)
+	Construct x xs == Construct y ys = (x == y) * (xs == ys)
 
 instance (Semigroup a, forall b . Semigroup b => Semigroup (t b)) => Semigroup (Construction t a) where
-	Construction x xs + Construction y ys = Construction (x + y) $ xs + ys
+	Construct x xs + Construct y ys = Construct (x + y) $ xs + ys
 
 instance (Monoid a, forall b . Semigroup b => Monoid (t b)) => Monoid (Construction t a) where
-	zero = Construction zero zero
+	zero = Construct zero zero
 
 deconstruct :: Construction t a -> (t :. Construction t) a
-deconstruct (Construction _ xs) = xs
+deconstruct (Construct _ xs) = xs
 
 coiterate :: Covariant t => a |-> t -> a |-> Construction t
-coiterate coalgebra x = Construction x $ coiterate coalgebra <$> coalgebra x
+coiterate coalgebra x = Construct x $ coiterate coalgebra <$> coalgebra x
 
 section :: Comonad t => t ~> Construction t
-section as = Construction (extract as) $ extend section as
+section as = Construct (extract as) $ extend section as
 
 instance (Covariant t, Covariant u) => Covariant (TU Covariant Covariant u (Construction t)) where
 	f <$> TU g = TU $ f <$$> g
 
 instance (Avoidable t, Pointable u) => Pointable (TU Covariant Covariant u (Construction t)) where
-	point x = TU . point . Construction x $ empty
+	point x = TU . point . Construct x $ empty
 
 instance (Applicative t, Applicative u) => Applicative (TU Covariant Covariant u (Construction t)) where
 	TU f <*> TU x = TU $ f <**> x