+++ CHANGELOG.md
@@ -215,3 +215,4 @@
 * Define `Rose` tree structure
 * Replace `UT` joint scheme with `TU` in data structures
 * Change `sub` method of `Substructure` - it always returns `Tagged` value
+* Rename `Construction` constructor of `Consturction` with `Construct`